requirements.txt
huggingface_hub>=0.30
python-dotenv